Refactored <FoldOut> component
